pyproject.toml
[project]
name = "lexikon"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "spacy>=3",
    "google-cloud-translate>=3",
]

[dependency-groups]
dev = [
    "pytest>=7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/lexikon"]
